+++ gdx-fireapp-core/src/mk/gdx/firebase/helpers/JavaCoreClassDetector.java
@@ -14,23 +14,22 @@
  * limitations under the License.
  */
 
-package mk.gdx.firebase.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+package mk.gdx.firebase.helpers;
 
 /**
- * Keeps information about Nested generic types ex. {@code doSomething(new Callback<List<User>>)}
- *
- * @deprecated You should use {@link MapConversion} instead.
+ * Gives information about type origin.
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-@Deprecated
-public @interface NestedGenericType
+public class JavaCoreClassDetector
 {
-    Class<?> value();
+
+    /**
+     * Check if given type package name starts with 'java'.
+     *
+     * @param type Type to examination, not null
+     * @return True if type is core java class.
+     */
+    public static boolean isJavaCoreClass(Class<?> type)
+    {
+        return type.getPackage().getName().startsWith("java");
+    }
 }